requirements.txt
# Requirements
#  Finrl
alpaca_trade_api>=2.1.0
ccxt>=1.90.41


pygame>=2.5.0

# elegantrl
gputil>=1.4.0
gym>=0.26.0
jqdatasdk>=1.8.10
lz4>=4.0.1
# plot
matplotlib>=3.5.2

numpy>=1.23.1
pandas>=1.4.3

# market data & paper trading API
pandas_market_calendars
pandas_market_calendars # calendars

# hooks
pre-commit>=2.19.0

# testing requirements
pytest>=7.1.2
ray[default]>=1.13.0

# Model Building Requirements
scikit-learn>=1.1.1

# packaging
setuptools>=63.1.0
stable-baselines3[extra]>=1.5.0
stockstats>=0.4.1
tensorboard
vectorbt[full]>=0.23.3
wheel>=0.37.1
wrds>=3.1.1
yfinance>=0.1.72

#  Autoformer
scikit-learn
torchvision
numpy
matplotlib
reformer_pytorch
